+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - POTENTIALLY BREAKING: Fixed a typing bug where the `totalSupply` field in an `NftContract` should have type `string` instead of `number`.
 
+### Minor Changes
+
+- Added support for the `erc20` token type and pagination for `CoreNamespace.getTokenBalances()`.
+
 ## 2.0.4
 
 ### Minor Changes